pyproject.toml
[project]
name = "base-cracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Angus Jull",email = "angusjull1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "pytest (>=9.0.1,<10.0.0)",
    "google-adk (==1.17.0)",
    "litellm (==1.79.0)",
    "python-dotenv (==1.2.1)",
    "google-genai (>=1.50.1,<2.0.0)",
    "chromadb (>=1.3.4,<2.0.0)",
    "langchain (>=1.0.7,<2.0.0)",
    "langchain-ollama (>=1.0.0,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "langchain-chroma (>=1.0.0,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]

[tool.poetry]
packages = [{include = "base_cracker", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
